chore: 🤖 Bump polmesh-sdk to `24.2.1`
