+++ sdk-docs-typedoc.json
@@ -7,7 +7,7 @@
     "src/generated/types.ts",
     "src/types/index.ts",
     "src/types/utils",
-    "src/utils/index.ts",
+    "src/entities/types.ts",
     "src/base"
   ],
   "entryPointStrategy": "expand",
