+++ src/api/entities/Identity/index.ts
@@ -26,16 +26,6 @@ export interface UniqueIdentifiers {
  */
 export class Identity extends PublicIdentity {
   /**
-   * @hidden
-   * Checks if a value is of type {@link UniqueIdentifiers}
-   */
-  public static override isUniqueIdentifiers(identifier: unknown): identifier is UniqueIdentifiers {
-    const { did } = identifier as UniqueIdentifiers;
-
-    return typeof did === 'string';
-  }
-
-  /**
    * Get Confidential Transactions affirmations involving this identity
    *
    * @note supports pagination
