+++ src/types/index.ts
@@ -20,7 +20,7 @@ import {
 import BigNumber from 'bignumber.js';
 
 import { CountryCode, ModuleName, TxTag, TxTags } from '~/generated/types';
-import { CheckPermissionsResult } from '~/types/internal';
+import { CheckPermissionsResult } from '~/types';
 
 export { EventRecord } from '@polkadot/types/interfaces';
 export { ConnectParams } from '~/api/client/Polymesh';
